Expose the RouteConfig interface.
I am trying to build a helper function to register schemas. Having access to the RouteConfig interface would be helpful here, because I could build a function that takes in a RouteConfig.

Instead of passing the registry around everywhere, I'd like to expose my helper function to the code that needs it.